+++ src/state_flow/assertions/matcher_combinators.clj
@@ -1,6 +1,6 @@
 (ns state-flow.assertions.matcher-combinators
   (:require [cats.core :as m]
-            [matcher-combinators.core :as matcher-combinators]
+            [matcher-combinators.standalone :as matcher-combinators]
             [matcher-combinators.test] ;; to register clojure.test assert-expr for `match?`
             [state-flow.core :as core]
             [state-flow.probe :as probe]
@@ -9,21 +9,28 @@
 (defn ^:private match-probe
   "Internal use only.
 
-  Returns the right value returned by probe/probe."
-  [state matcher params]
-  (m/fmap (comp :value last)
-          (probe/probe state
-                       #(matcher-combinators/match? (matcher-combinators/match matcher %))
-                       params)))
+  Returns the result of calling probe/probe with a function that
+  uses matcher-combinators to match the actual value.
+
+  args:
+  - step is a monad which should produce the actual value
+  - expected is any valid first argument to `matcher-combinators/match?`
+  - params are passed directly to probe "
+  [step expected params]
+  (probe/probe step
+               (fn [actual] (matcher-combinators/match? expected actual))
+               params))
 
 (defmacro match?
-  "Builds a state-flow assertion using matcher-combinators.
+  "Builds a state-flow step which uses matcher-combinators to make an
+  assertion.
 
-  - expected can be a literal value or a matcher-combinators matcher
-  - actual can be a literal value, a primitive step, or a flow
-  - params is an optional map supporting:
-    - :times-to-try optional, default 1
-    - :sleep-time   optional, millis to wait between tries, default 200
+  `expected` can be a literal value or a matcher-combinators matcher
+  `actual` can be a literal value, a primitive step, or a flow
+  `params` are optional keyword-style args, supporting:
+
+    :times-to-try optional, default 1
+    :sleep-time   optional, millis to wait between tries, default 200
 
   Given (= :times-to-try 1), match? will evaluate `actual` just once.
 
@@ -31,17 +38,20 @@
   retry up to :times-to-try times, waiting :sleep-time between each try,
   and stopping when `actual` produces a value that matches `expected`.
 
-  See `state-flow.probe/probe` for more info"
-  [expected actual & [{:keys [times-to-try]
-                       :or {times-to-try 1}
-                       :as params}]]
+  Returns a map (in the left value) with information about the success
+  or failure of the match, the details of which are used internally by
+  state-flow and subject to change."
+  [expected actual & [{:keys [times-to-try
+                              sleep-time]
+                       :as   params}]]
    ;; description is here to support the
    ;; deprecated cljtest/match? fn.  Undecided
    ;; whether we want to make it part of the API.
    ;; caller-meta is definitely not part of the API.
-  (let [params* (merge {:description "match?"
-                        :caller-meta (meta &form)
-                        :times-to-try 1}
+  (let [params* (merge {:description  "match?"
+                        :caller-meta  (meta &form)
+                        :times-to-try 1
+                        :sleep-time   probe/default-sleep-time}
                        params)]
     (core/flow*
      {:description (:description params*)
@@ -51,11 +61,21 @@
       ;; information from core/current-description.
      `(m/do-let
        [flow-desc# (core/current-description)
-        actual#    (if (> (:times-to-try ~params*) 1)
-                     (#'match-probe (state/ensure-step ~actual) ~expected ~params*)
-                     (state/ensure-step ~actual))]
-        ;; TODO: (dchelimsky, 2020-02-11) we plan to decouple
-        ;; assertions from reporting in a future release. Remove this
-        ;; next line when that happens.
+        probe-res# (#'match-probe (state/ensure-step ~actual) ~expected ~params*)
+        :let [actual# (-> probe-res# last :value)
+              report# (assoc (matcher-combinators/match ~expected actual#)
+                             :match/expected     ~expected
+                             :match/actual       actual#
+                             :probe/results      probe-res#
+                             :probe/sleep-time   ~(:sleep-time params*)
+                             :probe/times-to-try ~(:times-to-try params*))]]
+       ;; TODO: (dchelimsky, 2020-02-11) we plan to decouple
+       ;; assertions from reporting in a future release. Remove this
+       ;; next line when that happens.
        (state/wrap-fn #(~'clojure.test/testing flow-desc# (~'clojure.test/is (~'match? ~expected actual#))))
-       (state/return actual#)))))
+       (state/return report#)))))
+
+(defn report->actual
+  "Returns the actual value from the report returned by `match?`."
+  [report]
+  (:match/actual report))